dom/base/test/test_bug656283.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=656283
-->
<head>
  <title>Test for Bug 656283</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="test()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=656283">Mozilla Bug 656283</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 656283 **/

SimpleTest.waitForExplicitFinish();

var ifr;
function test() {
  // First a test with data documents.
  var d = document.implementation.createHTMLDocument("");
  is(d.activeElement, d.body, "Active element should be body by default! (1)");
 

  ifr = document.getElementById("ifr");
  ifr.onload = test2;
  ifr.src = "data:text/html,1";
}

var firstDoc;
function test2() {
  firstDoc = ifr.contentDocument;
  is(firstDoc.activeElement, firstDoc.body,
     "Active element should be body by default! (2)");
  ifr.onload = test3;
  ifr.src = "data:text/html,<input>";
}

function test3() {
  ifr.contentDocument.getElementsByTagName("input")[0].focus();
  is(firstDoc.activeElement, firstDoc.body,
     "Active element should be body by default! (3)");
  ifr.parentNode.removeChild(ifr);
  is(firstDoc.activeElement, firstDoc.body,
     "Active element should be body by default! (4)");
  SimpleTest.finish();
}

</script>
</pre>
<iframe id="ifr"></irame>
</body>
</html>
